+++ bin/adminer4.1.0/config.php
@@ -1,10 +1,17 @@
 <?php
 
-global $mysqlPort, $mysqlRootUser, $mysqlRootPwd, $mariadbPort, $mariadbRootUser, $mariadbRootPwd;
+global $mysqlPort, $mysqlRootUser, $mysqlRootPwd,
+$mariadbPort, $mariadbRootUser, $mariadbRootPwd,
+$postgresqlPort, $postgresqlRootUser, $postgresqlRootPwd;
 
 $mysqlPort = 3306;
 $mysqlRootUser = 'root';
 $mysqlRootPwd = '';
+
 $mariadbPort = 3307;
 $mariadbRootUser = 'root';
 $mariadbRootPwd = '';
+
+$postgresqlPort = 5432;
+$postgresqlRootUser = 'postgres';
+$postgresqlRootPwd = '';